refactor apphun API
